Update 75 Kotlin files (writeChar,writeBytes,Player,get) (+901 −586 lines) - Update buffer/src/main/kotlin/world/gregs/voidps/buffer/write/ArrayWriter.kt - Update buffer/src/main/kotlin/world/gregs/voidps/buffer/write/BufferWriter.kt - Update buffer/src/main/kotlin/world/gregs/voidps/buffer/write/Writer.kt (writeChar,writeBytes) - Update engine/src/main/kotlin/world/gregs/voidps/engine/client/EncodeExtensions.kt (Player) - Update engine/src/main/kotlin/world/gregs/voidps/engine/client/ui/Interfaces.kt (get,contains,refresh,sendPosition) - Update engine/src/main/kotlin/world/gregs/voidps/engine/client/update/npc/NPCUpdateTask.kt - Update engine/src/main/kotlin/world/gregs/voidps/engine/client/update/player/PlayerUpdateTask.kt - Update engine/src/main/kotlin/world/gregs/voidps/engine/client/variable/VariableBits.kt - Update engine/src/main/kotlin/world/gregs/voidps/engine/entity/character/Character.kt (clearWalkTrigger,walkTrigger,gfx,face) - Update engine/src/main/kotlin/world/gregs/voidps/engine/entity/character/Visuals.kt (Character) - Update engine/src/main/kotlin/world/gregs/voidps/engine/entity/character/npc/NPCSpawns.kt (loadNpcSpawns) - Update engine/src/main/kotlin/world/gregs/voidps/engine/entity/character/npc/NPCVisuals.kt - Update engine/src/main/kotlin/world/gregs/voidps/engine/entity/character/player/PlayerVisuals.kt (Player) - Update engine/src/main/kotlin/world/gregs/voidps/engine/entity/character/player/equip/BodyParts.kt - Update game/src/main/kotlin/GameTick.kt - Update game/src/main/kotlin/content/bot/behaviour/action/BotInterfaceOption.kt - Update game/src/main/kotlin/content/bot/behaviour/action/BotSpecAttack.kt - Update game/src/main/kotlin/content/bot/behaviour/action/BotSwitchSetup.kt - Update game/src/main/kotlin/content/entity/Examines.kt - Update network/src/main/kotlin/world/gregs/voidps/network/login/Protocol.kt (val) - Update network/src/main/kotlin/world/gregs/voidps/network/login/protocol/JagExtensions.kt (ByteReadChannel,ByteWriteChannel,Source) - Update network/src/main/kotlin/world/gregs/voidps/network/login/protocol/decode/ChatSetModeDecoder.kt - Update network/src/main/kotlin/world/gregs/voidps/network/login/protocol/decode/ChatTypeDecoder.kt - Update network/src/main/kotlin/world/gregs/voidps/network/login/protocol/decode/ClanChatRankDecoder.kt - Update network/src/main/kotlin/world/gregs/voidps/network/login/protocol/decode/ClientDetailOptionsStatusDecoder.kt - Update network/src/main/kotlin/world/gregs/voidps/network/login/protocol/decode/ConsoleCommandDecoder.kt - Update network/src/main/kotlin/world/gregs/voidps/network/login/protocol/decode/DialogueContinueDecoder.kt - Update network/src/main/kotlin/world/gregs/voidps/network/login/protocol/decode/FloorItemExamineDecoder.kt - Update network/src/main/kotlin/world/gregs/voidps/network/login/protocol/decode/FloorItemOption1Decoder.kt - Update network/src/main/kotlin/world/gregs/voidps/network/login/protocol/decode/FloorItemOption2Decoder.kt - Update network/src/main/kotlin/world/gregs/voidps/network/login/protocol/decode/FloorItemOption3Decoder.kt - Update network/src/main/kotlin/world/gregs/voidps/network/login/protocol/decode/FloorItemOption4Decoder.kt - Update network/src/main/kotlin/world/gregs/voidps/network/login/protocol/decode/FloorItemOption5Decoder.kt - Update network/src/main/kotlin/world/gregs/voidps/network/login/protocol/decode/InterfaceOnFloorItemDecoder.kt - Update network/src/main/kotlin/world/gregs/voidps/network/login/protocol/decode/InterfaceOnInterfaceDecoder.kt - Update network/src/main/kotlin/world/gregs/voidps/network/login/protocol/decode/InterfaceOnNpcDecoder.kt - Update network/src/main/kotlin/world/gregs/voidps/network/login/protocol/decode/InterfaceOnObjectDecoder.kt - Update network/src/main/kotlin/world/gregs/voidps/network/login/protocol/decode/InterfaceOnPlayerDecoder.kt - Update network/src/main/kotlin/world/gregs/voidps/network/login/protocol/decode/InterfaceOptionDecoder.kt - Update network/src/main/kotlin/world/gregs/voidps/network/login/protocol/decode/InterfaceSwitchComponentsDecoder.kt - Update network/src/main/kotlin/world/gregs/voidps/network/login/protocol/decode/NPCExamineDecoder.kt - Update network/src/main/kotlin/world/gregs/voidps/network/login/protocol/decode/NPCOption1Decoder.kt - Update network/src/main/kotlin/world/gregs/voidps/network/login/protocol/decode/NPCOption2Decoder.kt - Update network/src/main/kotlin/world/gregs/voidps/network/login/protocol/decode/NPCOption3Decoder.kt - Update network/src/main/kotlin/world/gregs/voidps/network/login/protocol/decode/NPCOption4Decoder.kt - Update network/src/main/kotlin/world/gregs/voidps/network/login/protocol/decode/NPCOption5Decoder.kt - Update network/src/main/kotlin/world/gregs/voidps/network/login/protocol/decode/ObjectExamineDecoder.kt - Update network/src/main/kotlin/world/gregs/voidps/network/login/protocol/decode/ObjectOption1Decoder.kt - Update network/src/main/kotlin/world/gregs/voidps/network/login/protocol/decode/PlayerOption10Decoder.kt - Update network/src/main/kotlin/world/gregs/voidps/network/login/protocol/decode/PlayerOption1Decoder.kt - Update network/src/main/kotlin/world/gregs/voidps/network/login/protocol/decode/PlayerOption3Decoder.kt - Update network/src/main/kotlin/world/gregs/voidps/network/login/protocol/decode/PlayerOption4Decoder.kt - Update network/src/main/kotlin/world/gregs/voidps/network/login/protocol/decode/PlayerOption5Decoder.kt - Update network/src/main/kotlin/world/gregs/voidps/network/login/protocol/decode/PlayerOption7Decoder.kt - Update network/src/main/kotlin/world/gregs/voidps/network/login/protocol/decode/PlayerOption9Decoder.kt - Update network/src/main/kotlin/world/gregs/voidps/network/login/protocol/decode/WalkMapDecoder.kt - Update network/src/main/kotlin/world/gregs/voidps/network/login/protocol/decode/WalkMiniMapDecoder.kt - Update network/src/main/kotlin/world/gregs/voidps/network/login/protocol/decode/WorldMapClickDecoder.kt - Update network/src/main/kotlin/world/gregs/voidps/network/login/protocol/encode/ChatStatusEncoder.kt - Update network/src/main/kotlin/world/gregs/voidps/network/login/protocol/encode/InterfaceEncoders.kt (Client) - Update network/src/main/kotlin/world/gregs/voidps/network/login/protocol/encode/NPCUpdateEncoder.kt - Update network/src/main/kotlin/world/gregs/voidps/network/login/protocol/encode/PlayerUpdateEncoder.kt - Update network/src/main/kotlin/world/gregs/voidps/network/login/protocol/encode/RunEnergyEncoder.kt - Update network/src/main/kotlin/world/gregs/voidps/network/login/protocol/encode/ScriptEncoder.kt - Update network/src/main/kotlin/world/gregs/voidps/network/login/protocol/encode/TextTileEncoder.kt - Update network/src/main/kotlin/world/gregs/voidps/network/login/protocol/encode/VarbitEncoder.kt - Update network/src/main/kotlin/world/gregs/voidps/network/login/protocol/encode/VarcEncoder.kt - Update network/src/main/kotlin/world/gregs/voidps/network/login/protocol/encode/VarpEncoder.kt - Update network/src/main/kotlin/world/gregs/voidps/network/login/protocol/encode/WeightEncoder.kt - Update network/src/main/kotlin/world/gregs/voidps/network/login/protocol/encode/ZoneEncoders.kt - Update network/src/main/kotlin/world/gregs/voidps/network/login/protocol/visual/PlayerVisuals.kt - Update network/src/main/kotlin/world/gregs/voidps/network/login/protocol/visual/Visuals.kt - Update network/src/main/kotlin/world/gregs/voidps/network/login/protocol/visual/encode/npc/NPCHitsEncoder.kt - Update network/src/main/kotlin/world/gregs/voidps/network/login/protocol/visual/encode/player/PlayerHitsEncoder.kt - Update network/src/main/kotlin/world/gregs/voidps/network/login/protocol/visual/encode/player/TemporaryMoveTypeEncoder.kt

+++ network/src/main/kotlin/world/gregs/voidps/network/login/protocol/encode/VarbitEncoder.kt
@@ -1,5 +1,7 @@
 package world.gregs.voidps.network.login.protocol.encode
 
+import io.ktor.utils.io.writeInt
+
 import world.gregs.voidps.network.client.Client
 import world.gregs.voidps.network.login.Protocol.CLIENT_VARBIT
 import world.gregs.voidps.network.login.Protocol.CLIENT_VARBIT_LARGE
@@ -19,7 +21,7 @@ fun Client.sendVarbit(id: Int, value: Int) {
     } else {
         send(CLIENT_VARBIT_LARGE) {
             writeInt(value)
-            writeShort(id.toShort())
+            writeShort(id)
         }
     }
 }
